pyproject.toml
[tool.poetry]
name = "dcat-ap-no-validator-service"
version = "0.1.0"
description = "The service validates a graph against dcat-ap-no specification"
authors = ["Stig B. Dørmænen <stigbd@gmail.com>"]
license = "Apache-2.0"
readme = "README.md"
homepage = "https://github.com/Informasjonsforvaltning/dcat-ap-no-validator-service"
repository = "https://github.com/Informasjonsforvaltning/dcat-ap-no-validator-service"
classifiers=[
  "Programming Language :: Python :: 3",
  "License :: OSI Approved :: Apache Software License",
  "Operating System :: OS Independent",
]
[tool.poetry.dependencies]
python = "^3.9"
gunicorn = "^20.0.4"
rdflib = "^5.0.0"
aiohttp = "^3.7.2"
python-dotenv = "^0.15.0"

[tool.poetry.dev-dependencies]
pytest = "^6.1.1"
coverage = {extras = ["toml"], version = "^5.3"}
pytest-dotenv = "^0.5.2"
pytest-docker = "^0.10.1"
pytest-cov = "^2.10.1"
docker-compose = "^1.27.4"
black = "^20.8b1"
flake8 = "^3.8.4"
flake8-bandit = "^2.1.2"
flake8-black = "^0.2.1"
flake8-bugbear = "^20.1.4"
flake8-docstrings = "^1.5.0"
flake8-import-order = "^0.18.1"
pep8-naming = "^0.11.1"
safety = "^1.9.0"
mypy = "^0.790"
flake8-annotations = "^2.4.1"
codecov = "^2.1.10"
requests-mock = "^1.8.0"
pytest-mock = "^3.3.1"
pytest-aiohttp = "^0.3.0"

[tool.coverage.paths]
source = ["src", "*/site-packages"]

[tool.coverage.run]
branch = false
source = ["dcat_ap_no_validator_service"]
omit = ["*gunicorn_config.py"]

[tool.coverage.report]
show_missing = true
fail_under = 100

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
